pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meds-torch-data"
dynamic = ["version"]
authors = [
  {name="Matthew McDermott", email="mattmcdermott8@gmail.com"},
  {name="Nassim Oufattole", email="noufattole@gmail.com"},
  {name="Teya Bergamaschi", email="teyabergamaschi@gmail.com"},
  {name="Aleksia Kolo", email="aleksiakolo@gmail.com"},
]
description = "An efficient, flexible PyTorch dataset class for MEDS data."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pytest",
    "polars~=1.30.0",
    "nested_ragged_tensors>=0.1.0",
    "numpy",
    "torch",
    "meds~=0.4.0",
    "MEDS_transforms~=0.5.2",
    "hydra-core",
    "omegaconf",
    "meds_testing_helpers>=0.3.0"
]

[dependency-groups]
dev = ["pre-commit<4", "ruff", "pytest-cov", "hypothesis"]
benchmarks = ["ml-mixins[memtrackable]>=0.2", "rootutils"]
docs = [
  "mkdocs==1.6.1", "mkdocs-material==9.6.7", "mkdocstrings[python]==0.28.2", "mkdocstrings-shell~=1.0",
  "mkdocs-gen-files==0.5.0", "mkdocs-literate-nav==0.6.1", "mkdocs-section-index==0.3.9",
  "mkdocs-git-authors-plugin==0.9.2", "mkdocs-git-revision-date-localized-plugin==1.3.0", "markdown-callouts"
]

[tool.setuptools_scm]

[project.optional-dependencies]
parallelized = ["hydra-joblib-launcher~=1.2.0"]
lightning = ["lightning~=2.5.1"]

[project.scripts]
MTD_preprocess = "meds_torchdata.preprocessing.__main__:main"

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--doctest-modules",
  "--ignore=docs",
  "--ignore=AGENTS.md",
  "--ignore=CONTRIBUTORS.md",
  "--doctest-glob=*.md",
]
markers = [
  "parallelized: mark test as parallelized, requiring the optional dependency 'hydra-joblib-launcher'",
  "lightning: mark test as requiring the optional dependency 'lightning'",
]
doctest_optionflags = ["NORMALIZE_WHITESPACE", "ELLIPSIS"]

[project.entry-points.pytest11]
meds_torchdata = "meds_torchdata.pytest_plugin"

[project.entry-points."MEDS_transforms.stages"]
tokenization = "meds_torchdata.preprocessing.tokenization:main"
tensorization = "meds_torchdata.preprocessing.tensorization:main"

[project.urls]
Homepage = "https://github.com/mmcdermott/meds-torch-data"
Issues = "https://github.com/mmcdermott/meds-torch-data/issues"

[tool.ruff]
target-version = "py311"
line-length = 110
extend-exclude = ["docs/index.md"]

# Enable linters
[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "RUF", # Ruff-specific rules
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "N",   # pep8-naming
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__ files
"test_*" = ["DTZ001"]  # Ignore datetimez in test files

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["*MEDS*", "*NRT*", "*_SE_*", "*_SEM_*", "*_SM_*", "D", "L"]
